+++ cloudwatch.tf
@@ -5,8 +5,9 @@
  */
 
 resource "aws_cloudwatch_log_group" "waggledance_ecs" {
-  name = local.instance_alias
-  tags = var.tags
+  count = var.wd_instance_type == "ecs" ? 1 : 0
+  name  = local.instance_alias
+  tags  = var.tags
 }
 
 resource "aws_cloudwatch_dashboard" "apiary_federation" {