+++ sns.tf
@@ -1,3 +1,4 @@
 resource "aws_sns_topic" "apiary_federation_ops_sns" {
-  name = "${local.instance_alias}-operational-events"
+  count = var.wd_instance_type == "ecs" ? 1 : 0
+  name  = "${local.instance_alias}-operational-events"
 }